feat(ActionableList): add helper
